Add CSRF token verification middleware, update OAuth redirect URIs, and enhance email template links; include new user model fields for OAuth providers and two-factor authentication

+++ othello-fullstack/backend/services/cloudinary.service.js
@@ -11,10 +11,16 @@ const uploadOnCloudinary=async(localFilePath)=>{
         const response=await cloudinary.uploader.upload(localFilePath,{
            resource_type:"auto"
         })
+        if(!localFilePath.includes("https://ui-avatars.com/"))
+        {
         fs.unlinkSync(localFilePath);
+        }
         return response
     } catch (error) {
-        fs.unlinkSync(localFilePath);
+        if(!localFilePath.includes("https://ui-avatars.com/"))
+            {
+            fs.unlinkSync(localFilePath);
+            }
         return null;
     }
 }
